added creating new reservation

+++ MobilneTestiranjeBackend/src/main/java/com/example/mobilnetestiranjebackend/model/Accommodation.java
@@ -34,8 +34,8 @@ public class Accommodation {
     @ElementCollection
     private List<String> imagePaths;
 
-    private Integer minGuests;
-    private Integer maxGuests;
+    private Long minGuests;
+    private Long maxGuests;
 
     private AccommodationType accommodationType;
