+++ MobilneTestiranjeBackend/src/main/java/com/example/mobilnetestiranjebackend/repositories/OwnerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OwnerRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
 
-    Optional<Owner> findOwnerById(Integer ownerId);
+    Optional<Owner> findOwnerById(Long ownerId);
 }
